Получил ключь задачи

+++ finder/models.py
@@ -8,4 +8,4 @@ class Remains(models.Model):
     title = models.TextField(null=True)
     base_unit = models.CharField(max_length=10, null=True)
     project = models.CharField(max_length=30, null=True)
-    quantity = models.DecimalField(blank=True, null=True, max_digits=15, decimal_places=2)
+    quantity = models.FloatField(blank=True, null=True)
